pyproject.toml
[tool.poetry]
name = "wallet"
version = "0.1.0"
description = ""
authors = ["nikitagorojanin <ngorojanin@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.5"
django = "^5.1"
environs = "^11.0.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
mysqlclient = "^2.2.4"
drf-spectacular = "^0.27.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"
black = "^24.0.0"
mypy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
